build(cask): add essential packages for macOS desktop usage
Docker Desktop is only included because the necessary virtualysation
tools it ships with, which are necessary even thouh its primarily used
though CLI, and the GUI is unused. Docker Desktop (GUI) thus needs to be
running to expose the docker daemon socket.

DevPod Desktop (GUI) is included for a similar purpose: to enable the
CLI though the GUI button, since the `devpod` CLI is not directly
available as a `brew` formula.