feat(aichat): handle different locations on different systems
Config with same content is installed on different locations on darwin
and linux. For each config file, only a single file is maintained under
`.chezmoitemplates`, which is included as a resolved template in each
one of the different location files. The config for a given system is
ignored on different systems, so that it is only applied on its system.

docs: https://www.chezmoi.io/user-guide/manage-machine-to-machine-differences/#handle-different-file-locations-on-different-systems-with-the-same-contents