feat(chezmoi,direnv): add conventional dotenv file for template envvars
The new file is to contain envars used to configure the `chezmoi`
dynamic template initialization.

As a conventional dotenv file (`KEY=VALUE` pairs), it can be passed to
`devpod`, and additionally it is available by direnv in the home
environment (`~/.envrc`) so that `chezmoi init` uses its current config
without having to pass envvars explicitly.