feat(git): set default push mode to "simple"
"matichng" pushed all the local branches matching an upstream branch,
but "simple" only pushes the current branch and only if it matches an
upstream branch.

"simple" is much safer than "matching", and it is the default in Git
2.0. There is also "current", which pushes the current branch even if it
does not match any upstream branch.

With "simple", use `git push -u` to push local branches that do not
exist in the upstream yet.