feat(direnv): disable scrict_env and disable load_dotenv
Disable strict_env (which refuses to load if the .envrc script executes
with errors) for compatibility with devbox, which reported errors
because of a hook envvar.

Because of this, it is better to disable load_dotenv, because
failing to load .envrc falls back to loading .env, which can
be misleading.