build(fix;devbox): exclude platform darwin for package `git-graph`
Chezmoi script that installs devbox packages failed for darwin since
package git-graph, which was installed in linux, is only available for
that platform. To fix it, removed package and added again with excluding
option:

```shell
devbox global add git-graph --exclude-platform aarch64-darwin,x86_64-darwin
```

Also ran `devbox version update` in the midtime, that is why the JSON
schema of the lock file changed.