feat(amazon-q): enhance file exploration rules for clarity and enforcement
This commit improves the file exploration rules by:

- Converting preferences to explicit requirements
- Adding MUST directives for tool usage
- Including practical examples of fd and rg commands
- Explaining the performance and usability benefits

This improvement was created through an interactive Amazon Q session
where we identified ambiguity in the original rules and collaboratively
refined them to prevent misinterpretation.

Promp used after it used `find` and not `fd`:

```
> i was wondering why did you use `find` instead of `fd`
I used find instead of fd because I made a mistake in following the file
exploration rules that were provided in the context. The rules clearly
stated that fd is the preferred tool for finding files, but I
incorrectly used the standard find command instead.

This was an oversight on my part - I should have followed the specified
guidelines which recommend using fd for file discovery and rg (ripgrep)
for searching within files. I acknowledge this error and will be more
  careful to use the preferred tools in future file exploration tasks.

> the mistake was mine. help me improve the rules file so that there was
no change you wouldn't have followed it
I'll help you improve the rules file to make it clearer and more likely
to be followed. Here's a suggestion for enhancing the file exploration
rules:
```

Then told it to `chezmoi add` the modified file and `chezmoit git add`
it and `chezmoi git commit -- -m'...'`.