fix(zellij): make nvim available in the gitui tab process
Run as a bash command so that activation hooks are sourced
before running command gitui--specifically, the mise activation
hook, that makes its packages available in PATH, so nvim being
managed by mise is available for gitui.

This is of interest when e.g. <C-e> for editing a commit message
in the EDITOR / VISUAL (equal to nvim).