build(mise): migrate package `amazon-q` from backend `devbox` to `cargo`
This is required because nixhub has apparently stopped pushing new
versions of this package, or at least it is one month behind.

Since it is not a prebuild, like crates from crates.io, backend `cargo`
builds it from source (the GH repo), and is then a slow install, so put
it in mise environment `opt` instead of global.