feat(jj): re-enable colored output
The hidden change ID and commit ID characters, which are actually grey,
are not relevant at all for the UI workflow, since the brighter-colored
first characters (visible, not grey) is all the user really needs to
refer to an ID unambiguously, and that is what the UI outputs highlight.
It can even be a benefit to have "invisible" redundant ID characters,
from that point of view.